Centre navigation bar elements
Initially, I tried to make the navigation bar so that the elements enlarge when hovered over. The problem I faced was that hovering enlarged the element and its surroundings as well. The only way I was able to figure out how to fix this was to have the navigation bar elements displayed as block elements, which I didn’t want. Therefore, I returned the hover effect to changing the colour of the element. I also centred the navigation bar in the centre of the screen.