Add basic styles to section one of the main content
I’ve added some basic styles to the first section of the main content to give it a user-friendly look. By using a flex layout for both boxes, the images being side by side as well as the text being beside it makes it much easier, this is also the typical layout that restaurants use for their websites and I took inspiration from that.

I also limited the width of the box purely because I believe it looks more aesthetic.