Change the image of the hero banner
The reason behind this change was to better advertise the food of the restaurant. The hero banner should effectively represent the purpose of the restaurant in a way which catches the interest of the viewer and having a single dish as the image doesn't stand out compared to a table which is a melting pot of dishes from all around the world. Therefore, I changed it for the current image.