Make changes to subsection of the main
I realised that the Images weren't taking up the place of their container which is why my text and the images were overlapping when I changed the font size. The reason was because of the padding which I set for the images. I then sorted out the padding so that the images were taking up the correct parts of the container.