Add some style to the hero banner
Using the position property, as I did with the home page, I positioned the text of the hero banner to overlap the background image. The hero banner of this page is now a continuation of the one in the hero page, despite the minor adjustments to make it relevant to its specific webpage.