Style the sitewide sections for phone resolution
In this commit, I styled the sections that are universal to every webpage such as the header, branches section, and the footer.

Using just HTML and CSS, it is extremely hard to style the header navigation bar. That is because I cannot hide the nav elements and have them open as a dropdown. I styled the section anyway but it is for this reason that it looks unappealing.