Adjust a sections layout for phone resolution
In this commit, I adjusted the main section of the home page, excluding the branches section which is part of the main. One of the changes I made, which is unique to the phone resolution, is that I gave the images the block display so that they take up the full width of the screen. This is typical for many websites to do as it is visually more appealing for smaller screens to have larger images as it is harder to view on the small screen.